Replace docker push/pull with the regsync command (#1)
**What did I want**

In order to support multi-arch image retagging, we need to stop doing `docker pull/push` because that only pulls down a single image. Also that takes longer than is necessary.

**What did I do?**

I converted this action into a Docker-image based action where we build a Docker image with `regsync` and the `docker-cli` built in. The `regsync` command is used to do the replication, which is much smarter and faster... and will replicate the manifest only, which is all we need in this case. This greatly speeds up the action, AND makes it multi-arch compatible.